Extract Dialogs to own files
prepare generic contact list
